[refactor] QRcode modal 수정
